requirements.txt
pandas~=2.2.1
plotly~=5.20.0
streamlit~=1.32.2
pillow~=10.2.0